refactor: separated controller & service integration test
