+++ api/src/test/java/com/thesurvey/api/controller/BaseControllerTest.java
@@ -3,9 +3,6 @@ package com.thesurvey.api.controller;
 
 import com.thesurvey.api.domain.EnumTypeEntity.Role;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -16,9 +13,6 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.ResultActions;
 
-@SpringBootTest
-@Transactional
-@AutoConfigureMockMvc
 public abstract class BaseControllerTest {
 
     String globalName = "test name";
